Omskriving til typescript/funksjonell component
