+++ src/sider/eu_eøs/saksbehandling/stegMap/sokkel_skip.js
@@ -46,7 +46,6 @@ class SokkelSkip extends Steg {
     this.komponent = VurderingSokkelSkip;
     this.samleRelevanteData = (_propsLight) => ({
       begrunnelser: _propsLight.begrunnelser.sokkel,
-      behandlingstema: _propsLight.behandlingstema,
       redigerbart: _propsLight.generiskStegRedigerbart,
     });
     this.beregnRelevantUI = (_propsLight) => {
