--- src/sider/eu_eøs/stegKomponenter/vurderingSokkelSkip.jsx
@@ -1,190 +1,0 @@
-import { connect } from "react-redux";
-import PT from "prop-types";
-import { Component } from "react";
-import MKV from "@navikt/melosys-kodeverk";
-import * as Nav from "../../../navFrontend";
-import * as KV from "../../../kodeverk";
-import * as MPT from "../../../proptypes";
-import * as Utils from "../../../utils";
-import * as Mui from "../../../felleskomponenter/ui";
-import SokkelSkipListe from "../../../felleskomponenter/sokkelskipliste";
-import {
-  lagVilkaar,
-  slettVilkar,
-  konverterAvklartfaktaTilStegData,
-  lagAvklartefaktaBegrunnelse,
-  lagAvklartfakta,
-} from "../../../felleskomponenter/stegvelger";
-import { hentFaktaVerdi } from "../../../domeneUtils";
-
-import { formSelectors } from "../../../ducks/form";
-import "./vurderingSokkelSkip.css";
-
-class VurderingSokkelSkip extends Component {
-  componentDidMount() {
-    const { tilstand, oppdaterData } = this.props;
-    const { sokkelSkipKonklusjon } = tilstand;
-
-    oppdaterData(
-      konverterAvklartfaktaTilStegData(KV.Koder.avklartefaktaKoder.ARBEID_SOKKEL_SKIP, sokkelSkipKonklusjon)
-    );
-  }
-
-  componentWillUnmount() {
-    this.props.slettData();
-  }
-
-  avklartefaktaEndret = (type, subjektID, verdi) => {
-    const { oppdaterData } = this.props;
-    oppdaterData(lagAvklartfakta(type, subjektID, verdi, null));
-  };
-
-  avklartefaktaBegrunnelseEndret = (type, subjektID, verdi) => {
-    const { oppdaterData } = this.props;
-    oppdaterData(lagAvklartefaktaBegrunnelse(type, subjektID, [verdi]));
-  };
-
-  konklusjonEndretHandler = (event) => {
-    const { value } = event.target;
-    const { oppdaterData, slettData } = this.props;
-    this.avklartefaktaEndret(KV.Koder.avklartefaktaKoder.ARBEID_SOKKEL_SKIP, null, value);
-
-    if (value === KV.Koder.VurderingSokkelSkipTyper.SOKKEL_NORSK) {
-      oppdaterData(lagVilkaar("art11_3A", true));
-    } else {
-      slettData(slettVilkar("art11_3A"));
-    }
-  };
-
-  render() {
-    // Merknad fra møte 12.12.18: Vi må huske å gå innom “vurdering antall land”
-    // dersom man har valgt “to sokler / skip i flere land” siden vi går inn i artikkel 13.
-    const {
-      bekreftOgFortsett,
-      behandlingstema,
-      tilstand,
-      begrunnelser,
-      redigerbart,
-      oppdaterData,
-      slettData,
-      maritimtArbeid,
-      tilbake,
-    } = this.props;
-
-    const {
-      sokkelEllerSkipListe,
-      sokkelSkipKonklusjon,
-      installasjonArbeidslandListe,
-      installasjonArbeidslandTypeListe,
-      arbeidslandListe,
-    } = tilstand;
-    const fakta = hentFaktaVerdi(sokkelSkipKonklusjon);
-
-    const { konklusjonEndretHandler } = this;
-    const { VurderingSokkelSkipTyper } = KV.Koder;
-    const { harAvklaring } = tilstand;
-    const harMaritimeArbeidUnikeNavn = Utils.erPropertyUnik(
-      maritimtArbeid,
-      (enkeltMaritimtArbeid) => enkeltMaritimtArbeid.enhetNavn
-    );
-    /* eslint-disable max-len */
-    return (
-      <div className="vurderingSokkelSkip">
-        <Nav.Typo.Innholdstittel className="stegvelgertittel">Vurdering av sokkel eller skip</Nav.Typo.Innholdstittel>
-        <SokkelSkipListe
-          sokkelEllerSkipListe={sokkelEllerSkipListe}
-          installasjonArbeidslandListe={installasjonArbeidslandListe}
-          installasjonArbeidslandTypeListe={installasjonArbeidslandTypeListe}
-          arbeidslandListe={arbeidslandListe}
-          maritimtArbeid={maritimtArbeid}
-          begrunnelser={begrunnelser}
-          redigerbart={redigerbart && harMaritimeArbeidUnikeNavn}
-          avklartefaktaEndretHandler={this.avklartefaktaEndret}
-          avklartefaktaBegrunnelserEndretHandler={this.avklartefaktaBegrunnelseEndret}
-          oppdaterData={oppdaterData}
-          slettData={slettData}
-        />
-        {maritimtArbeid.length === 0 && (
-          <div className="sokkelSkip__varsel">
-            <Nav.Alert variant="warning">Det er ikke registrert verken sokkel eller skip.</Nav.Alert>
-          </div>
-        )}
-        {!harMaritimeArbeidUnikeNavn && (
-          <div className="sokkelSkip__varsel">
-            <Nav.Alert variant="warning">Det er registrert flere maritime arbeid med samme navn.</Nav.Alert>
-          </div>
-        )}
-        <Nav.Fieldset legend="Hvordan arbeider søkeren:">
-          <Nav.Radio
-            name={KV.Koder.avklartefaktaKoder.ARBEID_SOKKEL_SKIP}
-            disabled={
-              !redigerbart || behandlingstema?.kode === MKV.Koder.behandlinger.behandlingstema.UTSENDT_ARBEIDSTAKER
-            }
-            onChange={konklusjonEndretHandler}
-            checked={fakta === VurderingSokkelSkipTyper.SOKKEL_NORSK}
-            value={VurderingSokkelSkipTyper.SOKKEL_NORSK}
-            label="På norsk sokkel eller innenfor norsk territorialfarvann (art. 11.3.a)"
-          />
-          <Nav.Radio
-            name={KV.Koder.avklartefaktaKoder.ARBEID_SOKKEL_SKIP}
-            disabled={!redigerbart}
-            onChange={konklusjonEndretHandler}
-            checked={fakta === VurderingSokkelSkipTyper.SKIP_ETT_LAND}
-            value={VurderingSokkelSkipTyper.SKIP_ETT_LAND}
-            label="På skip registrert i ett land"
-          />
-          <Nav.Radio
-            name={KV.Koder.avklartefaktaKoder.ARBEID_SOKKEL_SKIP}
-            disabled={!redigerbart}
-            onChange={konklusjonEndretHandler}
-            checked={fakta === VurderingSokkelSkipTyper.SOKKEL_UTLAND}
-            value={VurderingSokkelSkipTyper.SOKKEL_UTLAND}
-            label="Utsendt til sokkel eller til annet lands territorialfarvann (art. 12)"
-          />
-          <Nav.Radio
-            name={KV.Koder.avklartefaktaKoder.ARBEID_SOKKEL_SKIP}
-            disabled
-            onChange={konklusjonEndretHandler}
-            checked={fakta === VurderingSokkelSkipTyper.SOKKEL_ELLER_SKIP_FLERE_LAND}
-            value={VurderingSokkelSkipTyper.SOKKEL_ELLER_SKIP_FLERE_LAND}
-            label="To sokler / skip i flere land (art. 13)"
-          />
-        </Nav.Fieldset>
-        <Mui.StegKnapper
-          bekreftKnappProps={{
-            disabled: !(redigerbart && harAvklaring),
-            onClick: bekreftOgFortsett,
-          }}
-          tilbakeKnappProps={{
-            onClick: tilbake,
-            disabled: !redigerbart,
-          }}
-        />
-      </div>
-    );
-    /* eslint-enable max-len */
-  }
-}
-
-VurderingSokkelSkip.propTypes = {
-  begrunnelser: PT.arrayOf(MPT.Kodeverk).isRequired,
-  behandlingstema: MPT.Behandlingsresultat.isRequired,
-  bekreftOgFortsett: PT.func.isRequired,
-  maritimtArbeid: PT.array,
-  tilstand: PT.object,
-  redigerbart: PT.bool.isRequired,
-  oppdaterData: PT.func.isRequired,
-  slettData: PT.func.isRequired,
-  tilbake: PT.func.isRequired,
-};
-
-VurderingSokkelSkip.defaultProps = {
-  tilstand: {},
-  maritimtArbeid: [],
-};
-
-const mapStateToProps = (state) => ({
-  maritimtArbeid: formSelectors.MaritimtArbeidSelector(state),
-});
-
-export default connect(mapStateToProps)(VurderingSokkelSkip);
